refactor: api url 변경에 따라 channel, channels 컨트롤러 분리
